Fix issue #12, optimize activity start/end
Change signal chain for update when start/abort button is clicked in a building menu.
Now, instead of requesting update for everything in GUI, it requests update for Mercenaries Mode, fixing the bug and optimizing the procedure at the same time.
Also fix a trivial typo (updateEveryting() in main.qml).