Fix Map's landmarkInfo, remove unnecessary code, improve help.txt for VerInfoFileGen
Make Map's landmarkInfo always show inside the map, not half here, half there (as it used to be if the landmark was too close to the map border).
Remove Luck and SB handling in EquipmentsListDelegate in DoSt submenu (these bonuses are erased before going to this function and shouldn't be displayed in any way).
Make command in help.txt fully working.