Improve reports, make Hero wait with delivering reports when without signal, add QDataStream handling for Time
Add a new class, Report, as a general structured block of information from an EncounterReport (or some other report).
Replace EncounterReport with Report in Base.
Make Mission dtor public and add Mission deletion procedure in TimerAlarmsContainer::checkMissionAlarms(Time).
Generally improve that fnc so it longer automatically gives all the reports to the base - now it can also give that report to the mercenary and wait for reconnection.
Improve Hero so that it knows what to do with reports when disconnected (also change its save structure).
Add Base reports in Save.