Change way of planning missions
Now missions are planned fully at the start. This fixes the bug,
where after skipping a day game tried to plan next encounter
and the next encounter was to be executed on the same day
as the previous one, but day has already changed and
it could make the game plan next encounter on a day outside
of mission duration bounds, which resulted in the hero being
on the mission till the end of the world (mission end was planned
on -something day counting from the day of last encounter execution).