Add SignalTransmitter QML component
SignalTransmitter is a component that allows easy connection-establishing in QML object trees.
Each connection can be established without program knowing about the other object,
only the name of the receiver is important.
That way, signals can be for example connected to the grandgrandparents without having to forward the signal
through each parent.

Additionally improve includes in FPSCounter.