Improve integration between MainMenu and MainContent (proper game)
Remove dependency in BuildingsList so that now the component can be created without a working Base.
Add fullscreen mode to Settings and Credits to allow rendering a taller version of these components while in MainMenu, while preserving old positions when in the game.
Add clicks catcher in MainContent to make using MainMenu a safer experience.
Fix F4X::forceUIUpdate() to accomodate for the lack of updateEverythingInAMoment.